requirements.txt
gunicorn
torch
numpy
matplotlib
pandas
flask
torchvision
open_clip_torch
tqdm
transformers
wilds
pillow
keybert
datasets
yake
loguru
h5py
psutil
requests